fix: batch poster test (#1029)
